requirements.txt
kokoro>=0.9.4
soundfile
google-generativeai
google-genai
python-dotenv
moviepy
requests
selenium
